+++ Amazon.QLDB.Driver/driver/IQldbDriver.cs
@@ -46,7 +46,22 @@ namespace Amazon.QLDB.Driver
         /// <exception cref="TransactionAbortedException">Thrown if the Executor lambda calls <see cref="TransactionExecutor.Abort"/>.</exception>
         /// <exception cref="QldbDriverException">Thrown when called on a disposed instance.</exception>
         /// <exception cref="AmazonServiceException">Thrown when there is an error executing against QLDB.</exception>
+        [Obsolete("As of release 1.0, replaced by 'retryPolicy'. Will be removed in the next major release.")]
         void Execute(Action<TransactionExecutor> action, Action<int> retryAction);
+
+        /// <summary>
+        /// Execute the Executor lambda against QLDB within a transaction where no result is expected.
+        /// </summary>
+        ///
+        /// <param name="action">The Executor lambda with no return value representing the block of code to be executed within the transaction.
+        /// This cannot have any side effects as it may be invoked multiple times.</param>
+        /// <param name="retryPolicy">A <see cref="RetryPolicy"/> that overrides the RetryPolicy set when creating the driver. The given retry policy
+        /// will be used when retrying the transaction.</param>
+        ///
+        /// <exception cref="TransactionAbortedException">Thrown if the Executor lambda calls <see cref="TransactionExecutor.Abort"/>.</exception>
+        /// <exception cref="QldbDriverException">Thrown when called on a disposed instance.</exception>
+        /// <exception cref="AmazonServiceException">Thrown when there is an error executing against QLDB.</exception>
+        void Execute(Action<TransactionExecutor> action, RetryPolicy retryPolicy);
 
         /// <summary>
         /// Execute the Executor lambda against QLDB and retrieve the result within a transaction.
@@ -88,7 +103,30 @@ namespace Amazon.QLDB.Driver
         /// <exception cref="TransactionAbortedException">Thrown if the Executor lambda calls <see cref="TransactionExecutor.Abort"/>.</exception>
         /// <exception cref="QldbDriverException">Thrown when called on a disposed instance.</exception>
         /// <exception cref="AmazonServiceException">Thrown when there is an error executing against QLDB.</exception>
+        [Obsolete("As of release 1.0, replaced by 'retryPolicy'. Will be removed in the next major release.")]
         T Execute<T>(Func<TransactionExecutor, T> func, Action<int> retryAction);
+
+        /// <summary>
+        /// Execute the Executor lambda against QLDB and retrieve the result within a transaction.
+        /// </summary>
+        ///
+        /// <param name="func">The Executor lambda representing the block of code to be executed within the transaction. This cannot have any
+        /// side effects as it may be invoked multiple times, and the result cannot be trusted until the
+        /// transaction is committed.</param>
+        /// <param name="retryPolicy">A <see cref="RetryPolicy"/> that overrides the RetryPolicy set when creating the driver. The given retry policy
+        /// will be used when retrying the transaction.</param>
+        /// <typeparam name="T">The return type.</typeparam>
+        ///
+        /// <returns>The return value of executing the executor. Note that if you directly return a <see cref="IResult"/>, this will
+        /// be automatically buffered in memory before the implicit commit to allow reading, as the commit will close
+        /// any open results. Any other <see cref="IResult"/> instances created within the executor block will be
+        /// invalidated, including if the return value is an object which nests said <see cref="IResult"/> instances within it.
+        /// </returns>
+        ///
+        /// <exception cref="TransactionAbortedException">Thrown if the Executor lambda calls <see cref="TransactionExecutor.Abort"/>.</exception>
+        /// <exception cref="QldbDriverException">Thrown when called on a disposed instance.</exception>
+        /// <exception cref="AmazonServiceException">Thrown when there is an error executing against QLDB.</exception>
+        T Execute<T>(Func<TransactionExecutor, T> func, RetryPolicy retryPolicy);
 
         /// <summary>
         /// Retrieve the table names that are available within the ledger.